+++ mean_average_precision.egg-info/PKG-INFO
@@ -1,5 +1,5 @@
 Metadata-Version: 2.1
-Name: mAP
+Name: mean-average-precision
 Version: 0.0.1
 Summary: UNKNOWN
 Home-page: https://github.com/petersiemen/mAP
