+++ mean_average_precision.egg-info/requires.txt
