+++ mean_average_precision.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+mean_average_precision
