+++ setup.py
@@ -4,7 +4,7 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="mAP",
+    name="mean_average_precision",
     version="0.0.1",
     author="Peter Siemen",
     author_email="mail@petersiemen.net",
